test: add screenshot tests
