+++ packages/react-ui/.creevey/images/DropdownMenu/Functional tests/With Items And Icons Without Text Alignment/opened/chrome/opened.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:386115db28bb9fe7870160c8184203c09008314ed08bf07ae2eab7d34d2e9a2d
+size 7993
