+++ packages/react-ui/.creevey/images/DropdownMenu/Functional tests/With Items And Icons/opened/chrome/opened.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:73eb4604da63cc609cefbf9855110ba832993b9806c966f5ea280dc22ecfd732
+size 7989
