+++ packages/react-ui/.creevey/images/Kebab/Functional tests/With Items And Icons Without Text Alignment/opened/chrome/opened.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c6270533e9f98ee995dcbd69774aad703ea055f172023411451567af5e7443c2
+size 6622
