+++ packages/react-ui/.creevey/images/Kebab/Functional tests/With Items And Icons/opened/chrome/opened.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:752c78e5a11e063daadea9f116b2537a8e6a1d0b90c88c292b235161c63cf26b
+size 6613
