+++ packages/react-ui/.creevey/images/PopupMenu/With Items With Icons Without Text Alignment/opened/chrome/opened.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:563945f1e70bb49769f3ae4ec9ae4f33628a68b7c47ea7c95e24d96a748ccd6c
+size 10016
