+++ packages/react-ui/.creevey/images/PopupMenu/With Items With Icons/opened/chrome/opened.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a23a25b215a58129bd6df69a8ecfbca747aa0e3219c9da441240b3b0dfe9a0c9
+size 10009
