+++ packages/react-ui/.creevey/images/PopupMenu/with Items/opened/chrome/opened.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ae4b8300f4aeb894e57a325550a19fa80442e871c21f3674f5e8913d5fc3c07c
+size 8653
