+++ packages/react-ui/.creevey/images/TooltipMenu/Functional tests/With Items And Icons Without Text Alignment/opened/chrome/opened.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:db8d7ee81dd121884f6bde068ef7cbf1b850f23f74d4615633d966a728a2b45b
+size 9649
