+++ packages/react-ui/.creevey/images/TooltipMenu/Functional tests/With Items And Icons/opened/chrome/opened.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:02c3bd5df69b807e4f661f0d18740bb5edd787506e7e5dc96d64a1847b8e4f47
+size 9648
