+++ packages/react-ui/components/TooltipMenu/__stories__/TooltipMenu.stories.tsx
@@ -202,42 +202,3 @@ export const MenuWithoutAnimations = () => (
 );
 MenuWithoutAnimations.storyName = 'Menu without animations';
 MenuWithoutAnimations.parameters = { creevey: { skip: true } };
-
-export const WithItemsAndIcons = () => (
-  <div style={{ width: 200, textAlign: 'center' }}>
-    <TooltipMenu
-      caption={
-        <span style={{ display: 'inline-block' }} tabIndex={0}>
-          <MenuIcon size={32} />
-        </span>
-      }
-    >
-      <MenuHeader>MenuHeader</MenuHeader>
-      <MenuItem icon={<LightbulbIcon />}>MenuItem1</MenuItem>
-      <MenuItem icon={<LightbulbIcon />}>MenuItem2</MenuItem>
-      <MenuItem>MenuItem3</MenuItem>
-    </TooltipMenu>
-  </div>
-);
-
-export const WithItemsAndIconsWithoutTextAlignment = () => (
-  <div style={{ width: 200, textAlign: 'center' }}>
-    <TooltipMenu
-      caption={
-        <span style={{ display: 'inline-block' }} tabIndex={0}>
-          <MenuIcon size={32} />
-        </span>
-      }
-      enableTextAlignment={false}
-    >
-      <MenuHeader>MenuHeader</MenuHeader>
-      <MenuItem icon={<LightbulbIcon />}>MenuItem1</MenuItem>
-      <MenuItem icon={<LightbulbIcon />}>MenuItem2</MenuItem>
-      <MenuItem>MenuItem3</MenuItem>
-    </TooltipMenu>
-  </div>
-);
-
-WithItemsAndIconsWithoutTextAlignment.parameters = {
-  creevey: { skip: { 'themes dont affect logic': { in: /^(?!\bchrome\b)/ } } },
-};
